refactor: day, message vo로 리팩토링
