+++ src/main/java/com/backend/blooming/stamp/domain/Stamp.java
@@ -2,9 +2,9 @@ package com.backend.blooming.stamp.domain;
 
 import com.backend.blooming.common.entity.BaseTimeEntity;
 import com.backend.blooming.goal.domain.Goal;
-import com.backend.blooming.stamp.domain.exception.InvalidStampException;
 import com.backend.blooming.user.domain.User;
 import jakarta.persistence.Column;
+import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.ForeignKey;
@@ -19,9 +19,6 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
-
-import java.time.LocalDate;
-import java.time.temporal.ChronoUnit;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -45,11 +42,11 @@ public class Stamp extends BaseTimeEntity {
     @JoinColumn(name = "user_id", foreignKey = @ForeignKey(name = "fk_stamp_user"), nullable = false)
     private User user;
 
-    @Column(name = "stamp_day", nullable = false)
-    private int day;
+    @Embedded
+    private Day day;
 
-    @Column(columnDefinition = "text", nullable = false, length = STAMP_MESSAGE_MAXIMUM)
-    private String message;
+    @Embedded
+    private Message message;
 
     @Column(name = "is_deleted", nullable = false)
     private boolean deleted = false;
@@ -63,34 +60,7 @@ public class Stamp extends BaseTimeEntity {
     ) {
         this.goal = goal;
         this.user = user;
-        this.day = validateDay(goal, day);
-        this.message = validateMessage(message);
-    }
-
-    private int validateDay(final Goal goal, final int day) {
-        final long nowStampDay = ChronoUnit.DAYS.between(goal.getGoalTerm().getStartDate(), LocalDate.now()) + 1;
-
-        if (day > nowStampDay) {
-            throw new InvalidStampException.InvalidStampDayFuture();
-        }
-        if (day > goal.getGoalTerm().getDays()) {
-            throw new InvalidStampException.InvalidStampDay();
-        }
-        if (goal.getGoalTerm().getStartDate().isAfter(LocalDate.now())) {
-            throw new InvalidStampException.InvalidStampDay();
-        }
-
-        return day;
-    }
-
-    private String validateMessage(final String message) {
-        if (message == null || message.isEmpty()) {
-            throw new InvalidStampException.InvalidStampMessage();
-        }
-        if (message.length() > STAMP_MESSAGE_MAXIMUM) {
-            throw new InvalidStampException.InvalidStampMessage();
-        }
-
-        return message;
+        this.day = new Day(goal, day);
+        this.message = new Message(message);
     }
 }
